+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - Glow squids can now hypnotize you, displaying a hypnotizing shader that gradually gets stronger the longer you look at them
   - Glow squids will attract your cursor if they appear on your screen, this attraction getting stronger with the shader
   - Both cursor attraction and hypnotizing can be disabled with two new configuration options
+- Glow squids named "jeb_" now glow rainbow and have a rainbow hypnotizing pattern
 - Splashes and ripples (from rain or splashes) in warm oceans at night will now glow from glowing plankton
 -